binary_helpers::le

Trait LittleEndianInteger

pub trait LittleEndianInteger: Sized + Copy {
    const SIZE: usize;

    // Required methods
    fn from_le(source_bytes: &[u8]) -> Result<Self, BinaryError>;
    fn to_le(self, target_buffer: &mut [u8]) -> Result<(), BinaryError>;
}

A trait for types that can be serialized and deserialized in little-endian format. This trait is implemented for integer types like u16, u32, and u64.

Required Associated Constants

const SIZE: usize

The size of the type in bytes.

Required Methods

fn from_le(source_bytes: &[u8]) -> Result<Self, BinaryError>

Converts a slice of bytes in little-endian format to the target integer type.

Parameters

source_bytes: A slice of bytes to be converted. Must be exactly Self::SIZE bytes-long.

Returns

Ok(Self): The deserialized integer value.

Err(BinaryError): If the slice size does not match Self::SIZE.

fn to_le(self, target_buffer: &mut [u8]) -> Result<(), BinaryError>

Serializes the integer into a slice of bytes in little-endian format.

Parameters

target_buffer: A mutable slice where the serialized bytes will be written. Must be exactly Self::SIZE bytes-long.

Returns

Ok(()): If the serialization is successful.

Err(BinaryError): If the size of the output slice does not match Self::SIZE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.